refactor(test): amqp
