+++ test/test-amqp-bindings/bindings/amqp-bindings/password
