+++ test/test-amqp-bindings/bindings/rabbitmq-bindings/password
